Added TDD happy flow test case

+++ Engie.Powerplant.Lorenzo.Services/Models/FuelsModel.cs
@@ -8,6 +8,6 @@
 
         public decimal Co2 { get; set; }
 
-        public int Wind { get; set; }
+        public decimal Wind { get; set; }
     }
 }
